% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/den.R
\name{den}
\alias{den}
\title{Expnormal Distribution}
\usage{
den(x, a1, b1, a2, b2, c)
}
\arguments{
\item{x}{real argument}

\item{a1}{position parameter}

\item{b1}{positive scale parameter}

\item{a2}{position parameter}

\item{b2}{positive scale parameter}

\item{c}{semi-fraction parameter}
}
\value{
The function returns the value of the probability density function for the Expnormal distribution.
}
\description{
Density, distribution function, quantile function and random generation
for the Expnormal distribution with parameters a1, b1, a2, b2 and c.
}
\details{
Probability density function
see formula (2.1) in the article
Cumulative distribution function
see formula (2.3)
Quantile functon
see proposition (2.2)
Random number generator
see proposition (2.6)
}
\examples{
den(1,1,2,2,2,1)
pen(1,1,2,2,2,1)
qen(0.5,1,2,2,2,1)
ren(10,1,2,2,2,1)

}
\references{
{Sulewski, P. (2022). \emph{New Members of The Johnson Family of Probability Distributions:Properties and Application,} Accepted: February 2022. REVSTAT-Statistical Journal.}
}
\author{
Piotr Sulewski, \email{piotr.sulewski@upsl.edu.pl}, Pomeranian UNiwersity in Slupsk.
}
