% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peck.R
\name{peck}
\alias{peck}
\title{Easily Changeable Kurtosis Distribution}
\usage{
peck(x, a, p)
}
\arguments{
\item{x}{-a<x<a for -1<p<0 or -a<=x<=a for p>=1}

\item{a}{positive scale parameter}

\item{p}{shape parameter: p>-1}
}
\value{
The function returns the value of the cumulative distribution function for the Easily Changeable Kurtosis Distribution.
}
\description{
Density, distribution function, quantile function and random generation
for the Easily Changeable Kurtosis Distribution with parameters a and p.
}
\details{
Probability density function
see formula (1) or (3) in the article
Cumulative distribution function
see formula (4)
Quantile functon
see formula (20)
Random number generator
see formula (41)
}
\examples{
deck(1,2,3)
peck(1,2,3)
qeck(0.5,2,3)
reck(10,2,3)

}
\references{
{Sulewski, P. (2022). \emph{Easily Changeable Kurtosis Distribution.} Austrian Journal of Statistics 52, 1-24.}
}
\author{
Piotr Sulewski, \email{piotr.sulewski@upsl.edu.pl}, Pomeranian UNiwersity in Slupsk.
}
